Updated previews for 2048, minefield and fifteen. Made a single build script that builds all.

+++ tools/build/run-browser.sh
@@ -1,7 +1,8 @@
 #!/bin/bash
 
 # USAGE:
-#  build-browser.sh PACKAGENAME - builds package
+#  run-browser.sh PACKAGENAME      - builds and runs web
+#  run-browser.sh PACKAGENAME pkg  - builds package and runs web
 
 METAFILE="main.yaml"
 MASTERPREFIX="kote"
@@ -12,6 +13,12 @@ PROJECT="$1"
 if [ -z "$1" ]
   then
     PROJECT="allgames"
+fi
+
+MAKEPACKAGE="$2"
+if [ -z "$2" ]
+  then
+    MAKEPACKAGE=""
 fi
 
 
@@ -26,16 +33,22 @@ CWD=$(pwd)
 # 1. scaffold the app
 ./scaffold.sh $METAFILE $PROJECT
 
-# 2. browser build
-cd "$ROOTDIR"
-cordova build browser
+if [ -n "$MAKEPACKAGE" ]
+  then
+	# 2. browser build
+	cd "$ROOTDIR"
+	cordova build browser
 
-# prepare zip file
-rm -rf $TMPDIR/$TARGETNAME
-cp -r platforms/browser/www $TMPDIR/$TARGETNAME
-cd $TMPDIR
-mkdir -p $DISTDIR
-zip -r $DISTDIR/$TARGETNAME.zip $TARGETNAME
+	# prepare zip file
+	rm -rf $TMPDIR/$TARGETNAME
+	cp -r platforms/browser/www $TMPDIR/$TARGETNAME
+	cd $TMPDIR
+	mkdir -p $DISTDIR
+	zip -r $DISTDIR/$TARGETNAME.zip $TARGETNAME
+fi
+
+cd "$ROOTDIR"
+cordova run browser
 
 # change back to the original dir
 cd $CWD
